add_function_and_tests_averages_and_variances
